fix start command - currently you have to quit twice to kill the start command. add a finally clause to kill the child process - if there is another start already running, it still hangs. fix it to throw error right away
